requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.3.7
python-dotenv==1.0.0
jinja2==3.1.2
yagmail==0.15.293
schedule==1.2.0
lxml==4.9.3
markdown==3.5.1
pandas==2.1.3
numpy==1.25.2
webdriver-manager==4.0.1
httpx>=0.24.0
pytest 